Added models.py, email.py
