+++ app/__init__.py
@@ -26,4 +26,3 @@ def create_app(config_name):
     app.register_blueprint(main_blueprint)
 
     return app
-
