+++ app/main/forms.py
@@ -1,7 +1,6 @@
 from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField
 from wtforms.validators import DataRequired
-from . import main
 
 
 class NameForm(FlaskForm):
