--- main.py
@@ -1,83 +1,0 @@
-import os
-from threading import Thread
-
-from flask import Flask, render_template, session, redirect, url_for
-from flask_bootstrap import Bootstrap
-from flask_mail import Mail, Message
-from flask_migrate import Migrate
-from flask_moment import Moment
-from flask_sqlalchemy import SQLAlchemy
-from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField
-from wtforms.validators import DataRequired
-
-basedir = os.path.abspath(os.path.dirname(__file__))
-
-app = Flask(__name__)
-app.config["SECRET_KEY"] = os.getenv('SECRET_KEY')
-app.config["SQLALCHEMY_DATABASE_URI"] = 'sqlite:///' + os.path.join(basedir, 'data.sqlite')
-app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-app.config["MAIL_SERVER"] = 'smtp.googlemail.com'
-app.config["MAIL_PORT"] = 587
-app.config["MAIL_USE_TLS"] = True
-app.config["MAIL_USERNAME"] = os.getenv('MAIL_USERNAME')
-app.config["MAIL_PASSWORD"] = os.getenv('MAIL_PASSWORD')
-app.config["MAIL_SUBJECT_PREFIX"] = '[Social App]'
-app.config["SOCIAL_APP_MAIL_SENDER"] = f'Social App Admin {app.config["MAIL_USERNAME"]}'
-app.config["SOCIAL_APP_ADMIN"] = os.getenv("SOCIAL_APP_ADMIN")
-
-db = SQLAlchemy(app)
-bootstrap = Bootstrap(app)
-moment = Moment(app)
-migrate = Migrate(app, db)
-mail = Mail(app)
-
-
-class Role(db.Model):
-    __tablename__ = 'roles'
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(64), unique=True)
-    users = db.relationship('User', backref='role')
-
-    def __repr__(self):
-        return f'<Role {self.name}'
-
-
-class User(db.Model):
-    __tablename__ = 'users'
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(64), unique=True, index=True)
-    password = db.Column(db.String(64))
-    role_id = db.Column(db.Integer, db.ForeignKey('roles.id'))
-
-    def __repr__(self):
-        return f'<User {self.username}'
-
-
-@app.shell_context_processor
-def make_shell_contex():
-    return dict(db=db, User=User, Role=Role)
-
-
-
-
-
-def send_async_email(msg):
-    with app.app_context():
-        mail.send(msg)
-
-
-def send_email(to, subject, template, **kwargs):
-    msg = Message(app.config["MAIL_SUBJECT_PREFIX"] + " " + subject,
-                  sender=app.config["SOCIAL_APP_MAIL_SENDER"], recipients=[to])
-    msg.body = render_template(template + '.txt', **kwargs)
-    msg.html = render_template(template + '.html', **kwargs)
-    thr = Thread(target=send_async_email, args=[msg])
-    thr.start()
-    return thr
-
-
-db.create_all()
-
-if __name__ == '__main__':
-    app.run(port=8000, debug=True)
