+++ requirements.txt
@@ -1,0 +1,20 @@
+alembic==1.7.7
+blinker==1.4
+click==8.1.3
+dominate==2.6.0
+Flask==2.1.2
+Flask-Bootstrap==3.3.7.1
+Flask-Mail==0.9.1
+Flask-Migrate==3.1.0
+Flask-Moment==1.0.2
+Flask-SQLAlchemy==2.5.1
+Flask-WTF==1.0.1
+greenlet==1.1.2
+itsdangerous==2.1.2
+Jinja2==3.1.2
+Mako==1.2.0
+MarkupSafe==2.1.1
+SQLAlchemy==1.4.36
+visitor==0.1.3
+Werkzeug==2.1.2
+WTForms==3.0.1
